Refactor by using `index.ts` for modules
