+++ src/app.module.ts
@@ -5,11 +5,11 @@ import { ConfigModule } from '@nestjs/config';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { ProjectsModule } from '@projects/projects.module';
-import { RedisIoRedisModule } from './building-blocks/infra/redis-ioredis/redis-ioredis.module';
-import { RedisModule } from './building-blocks/infra/redis/redis.module';
-import { CachingModule } from './building-blocks/infra/caching/caching.module';
 import { DatabaseModule, DatabaseModuleSettings } from './building-blocks/infra/database';
 import { getDatabaseModuleSettings } from './typeorm.datasource';
+import { CachingModule } from '@building-blocks/infra/caching';
+import { RedisIoRedisModule } from './building-blocks/infra/redis-ioredis';
+import { RedisModule } from './building-blocks/infra/redis';
 
 const infrastructureModules = [
   DatabaseModule.register({
