+++ src/building-blocks/infra/caching/index.ts
@@ -1,0 +1,2 @@
+export * from "./caching.module";
+export * from "./caching.provider";
