+++ src/building-blocks/infra/redis-ioredis/index.ts
@@ -1,0 +1,2 @@
+export * from "./redis-ioredis.module";
+export * from "./redis-ioredis.service";
