+++ src/building-blocks/infra/redis/index.ts
@@ -1,0 +1,2 @@
+export * from "./redis.module";
+export * from "./redis.service";
