+++ src/projects/persistence/schemas/task.schema.ts
@@ -1,5 +1,5 @@
 import { Task } from "@projects/core/task";
-import { EntityBaseSchema } from "@src/building-blocks/infra/database/schemas/entity-base-schema";
+import { EntityBaseSchema } from "@building-blocks/infra/database/schemas/entity-base-schema";
 import { EntitySchema, EntitySchemaColumnOptions } from "typeorm";
 import { Project } from "@src/projects/core";
 import { ProjectSchema } from "./project.schema";
