+++ src/projects/use-cases/project/queries/find-by-id/find-by-id.handler.ts
@@ -3,14 +3,15 @@ import { FindByIdRequest } from "./find-by-id.request";
 import { FindByIdResponse } from "./find-by-id.response";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Repository } from "typeorm";
-import { CachingProvider } from "@src/building-blocks/infra/caching/caching.provider";
 import { Project } from "@src/projects/core";
+import { CachingProvider } from "@building-blocks/infra/caching";
 
 @QueryHandler(FindByIdRequest)
 export class FindByIdHandler implements IQueryHandler<FindByIdRequest, FindByIdResponse> {
 
     constructor(
-        @InjectRepository(Project) private readonly _projectRepository: Repository<Project>,
+        @InjectRepository(Project) 
+        private readonly _projectRepository: Repository<Project>,
         private readonly _cacheProvider: CachingProvider
     ){}
